// frontend/src/api/models/ChatFileResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type ChatFileResponse = {
    id: number;
    file_name: string;
    file_ext: string;
    file_size: number;
    content_type: (string | null);
    storage_type: string;
    url: string;
    md5: (string | null);
    parse_status: number;
    created_at: string;
};